// electron/src/sbg-api-client/interfaces/user.ts
export interface User {
    href?: string;
    city?: string;
    email?: string;
    phone?: string;
    state?: string;
    tags?: string[];
    country?: string;
    address?: string;
    username: string;
    zip_code?: string;
    last_name?: string;
    first_name?: string;
    affiliation?: string;
}